pyproject.toml
[tool.poetry]
name = "spotify-cover-maker"
version = "0.1.0"
description = ""
authors = ["Riley Flynn <riley@rileyflynn.me>"]
license = "MIT"
readme = "README.md"
packages = [{include = "spotify_cover_maker"}]

[tool.poetry.dependencies]
python = "^3.10"

CairoSVG = "^2.5.2"
Jinja2 = "^3.1.2"
PyYAML = "^6.0"
pydantic = "^1.10.4"
rich = "^12.6.0"
typer = {version = "^0.7.0", extras = ["all"]}

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^22.12.0"}
isort = "^5.11.4"
mypy = "^0.991"
types-PyYAML = "^6.0.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
